requirements.txt
Flask
flask_oauthlib
faunadb
requests
Pillow
rembg
gunicorn
